+++ main-service/src/main/java/ru/practicum/mainservice/event/dto/EventShortDtoOut.java
@@ -29,5 +29,5 @@ public class EventShortDtoOut {
     private Integer confirmedRequests;
 
     @Builder.Default
-    private Integer views = 0;
+    private Long views = 0L;
 }
